pratikproject/requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cx-Oracle==7.0.0
Django==2.1.4
django-filter==2.1.0
django-multiselectfield==0.1.8
djangorestframework==3.9.2
env==0.1.0
Faker==1.0.1
get==2018.11.19
httpie==1.0.2
idna==2.8
Markdown==3.0.1
post==2018.11.20
public==2018.11.20
Pygments==2.3.1
python-dateutil==2.7.5
pytz==2018.7
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
virtualenv==16.4.3
